settings: add gear icon to titlebar
As its been brought up in discord multiple times from new users, users
don't expect the plugin list to contain RuneLite's settings. This button
adds a shortcut to the RuneLite settings within the plugin settings
panel.

This is a different version that doesn't mess with NavigationButton.
It instead adds the same specialcallback to titletoolbarbuttons.

However it adds a slight discrepancy that the icon toggles, but the
normal settings icon does not toggle unless it gets clicked.