Add communication to launcher for splash screen
- Add property to connect to socket for launcher
- Send loading related messages to launcher

The passing of the socket is based on a property, similar to the
launcher version. This is because a normal use case wouldn't be by using it
with command line arguments and showing up in the --help. Also using a
property allows for less breakage in the future. For example, if the behavior
between communicating the launcher and client changes, then the launcher won't
break (launching a newer version of RuneLite with a now unknown
argument). No special module binding is needed either, since the
properties is just injected.